mudando o tipo de dado do preco, recebe double mas retorna BigDecimal

+++ src/main/java/com/bootcamp_w3_g3/model/dtos/request/ProdutoForm.java
@@ -24,7 +24,7 @@ public class ProdutoForm {
 
     private Integer codigoDoProduto;
     private String nome;
-    private BigDecimal preco;
+    private Double preco;
     private Double temperaturaIndicada;
 
     /**
